internal/weldr: make mocking API-used Solver easier
The way Weldr API currently mocks the depsolver is convoluted and
requires compiling a new Go binary from unit tests and dumping a JSON
file on the disk for it to read and return. Moreover, the implementation
heavily depends on the Solver internal implementation (IOW, how it
actually depsolves by calling an external tool). None of this should
matter for unit tests and it makes it unnecessarily hard to refactor
things in the Solver, which should be an internal implementation detail.

Refactor the API a bit, to make it easier to mock the Solver in unit
tests. This is the first step.

Signed-off-by: Tomáš Hozza <thozza@redhat.com>