inernal/weldr: rework setting repoID of depsolved packages in tests
Prepare for the switch of depsolver mock in test by defining the
expected repoIDs as variables and use those in tests, instead of literal
values. This will make it possible to set the repoIDs in the new
depsolve mock returned dnfjson.DepsolveResult.

While at it, simplify how repoIDs are set by setting them in the
function that returns mock depsolved package list.

Signed-off-by: Tomáš Hozza <thozza@redhat.com>