test: run older CIV tag on nightlies
CIV was recently updated [1] to cover changes in the upstream image
definitions [2] in Azure images.  The 'pr-456' tag is the latest that
does not include those changes.  Older versions of osbuild-composer are
being tested in nightlies for 9.7 and 10.1, so it should run with the
older image definitions that do not include the changes.

[1] https://github.com/osbuild/cloud-image-val/pull/457
[2] https://github.com/osbuild/images/pull/1814