worker: remove cloudwatch group option from aws.ec2 executor
Logs should be forwarded from the executor to the worker.

This facilitates isolating the executor from the internet. Cloudwatch
doesn't have a well-defined IP range, so it's tricky to update the
executor's security group in such a way to allow only cloudwatch
traffic. By forwarding any log traffic from the executor via the worker,
this problem is eliminated.

Note that this just prepares for the needed security group changes which
would isolate the executor from the internet.