internal/cloudapi/test: set pipeline_names in OSBuildJobResult
The unit tests were relying on the backward compatibility fallback
mechanism to set the pipeline_names in the OSBuildJobResult. This is not
the case any more for any of our infrastructure, because the value is
always set for each job.

In addition, check if pipeline_names are set in the OSBuildJobResult
when getting SBOMs. Previously, the code would end up dereferencing
`nil` value if the pipeline_names were not set and it would not be set
by the backward compatibility layer. Although this error should never
happen in reality, it is safer to check for it and return the proper
error.

Signed-off-by: Tomáš Hozza <thozza@redhat.com>