Test/api/gcp: set root partition size to at least 4 GiB
It turns out that gce-disk-expand is not able to expand the root
partition placed on an LVM LV. The Cloud API test case uses filesystem
customizations, which by default convert the partition table to use LVM.
It also turns out that the root partition in such case ends up being 3
GiB big. While this is enough for RHEL-9 and RHEL-10, it causes issues
for RHEL-8.10 test case, where the root partition ends up having only 20
KiB of free space left, which effectively makes the system unable to do
set user SSH keys, etc.

Let's workaround this for now by increasing the size of the root
partition and fix it later in osbuild/images by not using LVM for GCE
image when using FS customizations.

Signed-off-by: Tomáš Hozza <thozza@redhat.com>