schutzbot/deploy.sh: apply podman CNI network workaround on c9s
By updating the terraform ref and using newer c9s ci-runner, we started
to hit the issue that podman is falling back to using CNI network
backend when it detects existing container images in the local container
store. This is the situation for our ci-runner images, because we are
embedding container images at build time. The solution is to explicitly
set the podman network backend to 'netavark'.

Signed-off-by: Tomáš Hozza <thozza@redhat.com>