internal/cloudapi: check JobError in ComposeManifest endpoint
The API endpoint used to check only for errors when getting the osbuild
manifest from job dependencies results. However, if the actual manifest
serialization in the respective job failed, the endpoint would return no
error, but instead `nil` as the manifest value.

Make sure to check also the manifest generation job JobError, before
returning the manifest, and return the error instead if it is set.

Signed-off-by: Tomáš Hozza <thozza@redhat.com>