internal/cloudapi: add test for ComposeManifests API endpoint
Add a new unit test TestComposeManifests to verify the compose manifests
API endpoint responses when the manifest generation job succeeds and
fails. Due to the handling of the manifest generation job on the server
side, I had to modify the server code to allow overriding the
serializeManifest() function for testing purposes.

Note that the ComposeManifests endpoint incorrectly returns 'null' as a
manifest value in case the manifest generation job failed. This will be
fixed later.

Signed-off-by: Tomáš Hozza <thozza@redhat.com>