internal/cloudapi: ensure manifests are generated in multi-tenancy test
The multi-tenancy test case uses a simplified flow when all jobs are
finished with empty results. Subsequently, all relevant compose
endpoints are checked.

As a result, the manifest generation would always fail on empty
depsolved packages list, because the depsolve job results were
empty. This didn't used to cause any issues, because the compose
manifests API endpoint was not checking for manifest job errors and
would happily return a 'nil' manifest. Fixing the compose manifests
endpoint to not return 'nil', but an error instead breaks the unit
test.

Fix the unit test by ensuring that a reasonable minimal depsolve job
results are returned, so that a manifest can be successfully generated.

Signed-off-by: Tomáš Hozza <thozza@redhat.com>