test/api/all: set root partition size to at least 4 GiB
In [0] commit, we did the change for GCP image type only. We are hitting
the "No space left on device" error while building Azure and OCI image
types on RHEL-8.10. The images have 3 GiB root partition, which is not
big enough to accommodate all the installed content. We use CDN repos for
RHEL-8.10 testing, so most probably some recent batch update made the
difference and caused the failure.

Let's be consistent and make sure that the root partition min size is at
least 4 GiB for all image types that use filesystem customizations.

[0] 845f840529c80df2f00763dfc401fea95fc5e361

Signed-off-by: Tomáš Hozza <thozza@redhat.com>