internal/worker: adjust DepsolvedPackage struct to rpmmd.Package
Modify the worker.DepsolvedPackage DTO properties and types to match the
rpmmd.Package implementation. The backward compatibility between old
composer-worker and new composer-worker-server is achieved on the JSON
Unmarshaller level. This means that the server will be able to process
DepsolveJobResults from the old worker that uses DTO that matches the
old rpmmd.PackageSpec. We explicitly test for this scenario in the CI
where the new osbuild-composer is ran with old composer-worker.

The checksum is now represented as a struct, instead of a string value.
The DepsolvedPackageChecksum supports Unmarshalling from JSON string and
object representation. For backward compatibility reason, the struct is
still Marshalled as a string value, which is consistent with the old
implementation.

Signed-off-by: Tomáš Hozza <thozza@redhat.com>