osbuild-worker: parse pipeline names from manifest in new job type
When using the images library directly to generate a manifest, we have
access to the manifest source (pre-serialized manifest).  This object
provides a list of pipeline names, separated into build pipelines
(manifestSource.BuildPipelines()) and payload pipelines
(manifestSource.PayloadPipelines()).  The pipeline names are used in two
jobs:
1. In the osbuild job, the log in the result is unordered because
   pipeline results are returned in a map.  The list of pipeline names
   is used in this case to sort the job log in the order each pipeline
   was executed.
2. The koji finalize job uses the two pipeline name lists to separate
   the metadata from org.osbuild.rpm stages between build and payload,
   so that it can store package lists alongside the compose result
   separated into build packages and payload packages.

When using image-builder-cli, we don't have access to these pipeline
name lists directly.  To get them, let's parse the manifest into a
partial structure, identify all the pipeline names, and infer the build
pipeline names from the build property of the rest of the pipelines.

This approach violates one of our (soft) rules about job arguments and
job results: That they should always be considered opaque json blobs.
However, in the case of the manifest, I think it's acceptable,
especially since it only relies on the top-level structure of the
pipelines property, which we guarantee to be stable in osbuild.