internal/osbuildexecutor: forward progress to composer
Uses status scanner to monitor osbuild to generate progress
updates. Progress updates are sent every 30 seconds.

Only one level of subprogress needs to be supported. Because osbuild
manifests consist of pipelines and stages, more than 2 levels of
progress are not expected.

The result is still captured from stdout as before. The worker-executor
will stream the progress information over http, and save the result in a
separate file in the output archive.