cloudapi: add experimental ImageBuilderManifestGeneration option
Add an option to the cloudapi server config that enables generating
manifests with the ImageBuilderManifest job type.  Also will make it
possible to enable this using an environment variable.

Use the experimentalflags package from osbuild/images to read the
IMAGE_BUILDER_EXPERIMENTAL environment variable and check that
'image-builder-manifest-generation' is defined.  If it is, enable the
ImageBuilderManifestGeneration option in the composer server config.