test,CI: pass experimental options to systemd unit when deploying
In the deploy script, after installing all the test RPMs, add an
override to the unit file that sets the IMAGE_BUILDER_EXPERIMENTAL
environment variable for the osbuild-composer service with any values
from the global runner environment.

If image-builder-manifest-generation is enabled, install image-builder
from the repos.

Add a single API test job (api.sh) to the gitlab CI config that uses the
new experimental job type.

In the api.sh script, when image-builder-manifest-generation is enabled
in the test, verify it was actually used by searching for the warning
that's logged when the job is enqueued.