worker: support setting repositories for image-builder manifest
When calling image-builder manifest, we need to set repositories
configured in osbuild-composer for distros.  Without this, the
repository configs embedded in image-builder are used, but in the
service, we don't want to rely on those.  Instead, we need to use the
repositories configured in the service itself.

Write any configured repository configurations to a file in a temporary
directory under the datadir/repositories/ subdirectory.  This works with
image-builder's --data-dir option, which looks for a repository file
matching the target distribution name under the repositories/
subdirectory.

The on-disk repository format that rpmmd loads is slightly different
from rpmmd.RepoConfig.  The main difference is the single BaseURL in the
on-disk format vs an array of BaseURLs in rpmmd.RepoConfig.  Here, we
temporarily duplicate the on-disk format definition and convert to it
when saving the repository configuration file.  We'll have to
deduplicate this at a later stage.