cloudapi: enable librepo by default (and update tests)
This commit enables librepo by default in the cloudapi.

We have librepo in osbuild/images/ibcli/bib for a while now, c.f.
https://github.com/osbuild/image-builder-cli/pull/51
https://github.com/osbuild/bootc-image-builder/pull/786 (enable)
https://github.com/osbuild/bootc-image-builder/pull/840 (make default)

It is not configurable right now, I am not sure there is a reason
to keep libcurl around for manifest generation as its more fragile
but if it is desired I can look into that too. Having it configurable
would allow us to revert to the previous org.osbuild.curl in case
anything is wrong with the new downloader (and arguably it was
only excercised by a relatively small number of people compared
to how much the service is building).