osbuild-worker: support repo mtls configs in image-builder-manifest
Support setting repository MTLS configs in the new
image-builder-manifest job type.  This is equivalent to the same feature
in the traditional depsolve job.  MTLS configs for repositories are
defined in the worker by mapping a repository URL to the appropriate
cert values (CA, key, and cert).  The values are set on the repository
configurations matching the repo URL and passed to the depsolver.