internal/osbuildexecutor/ec2: return logs on build failures
If osbuild crashes on the worker-executor side its build endpoint will
return an error because it is unable to move the osbuild json result to
the output directory and create a tarball. This causes the status
scanner to error because it will receive invalid JSON.

Because the full stack trace coming from osbuild should be present in
the error that is handed back to the worker (so it can construct the job
result with all the error details), the osbuild output needs to be
fetched in case of build failures.

The status scanner will put the line it cannot scan into the error, but
it only reads up until the first newline. Thus osbuild's output cannot
be forwarded via the status scanner.