worker: link repos for image-builder into root of data-dir
The repositories are meant to be in a subdirectory under the data-dir,
but prior to v41 of image-builder [1], it would check the root of the
data-dir instead.  Link the repositories file from the new location
($datadir/repositories/$distro.json) to the old location
($datadir/$distro.json) and add a note to switch between the two paths
when we have a reliable and convenient way of getting the image-builder
version number.

[1] https://github.com/osbuild/image-builder-cli/releases/tag/v41