cloudapi: use PipelineNames from osbuild job result instead of args
When the PipelineNames are needed to process the osbuild job metadata,
don't use the property on the osbuild job args.  Instead, read them from
the osbuild job result.

The osbuild job args don't have the PipelineNames set when we're using
image-builder to generate the manifest, because the information is only
available after the manifest has been generated, before the osbuild job
is queued.  However, the PipelineNames should always be set on the
osbuild job result, regardless of the way that the manifest was created.