test/common: change expected RHSM fact value when using image-builder
When using image-builder to generate the manifest, the value of the RHSM
fact is "image-builder-cli".  We want to be able to keep the value
consistent for all images built in the service, but for now the value
isn't configurable, so let's modify the test until we can change it.

See https://issues.redhat.com/browse/HMS-9819