worker: support separate credentials for S3 uploads
Currently we use iam instance profiles for everything, these are useful
because no separate set of credentials need to be maintained. However,
they have a limitation when it comes to generating presigned urls for
objects in s3, instance profiles only allow for urls that are valid for
6 hours. Using a separate set of credentials just for s3 allows for
generating urls that are valid for 7 days.