osbuild-worker-executor: ignore item order in tarball in test
In TestBuildIntegration(), the contents of the output tarball
(`tar -tf`) are compared with an expected list of files.  The check is
done by direct string match with the output.  However, the order of
elements in the archive is not guaranteed to be stable.  Since the file
order isn't significant for the test or the feature, compare the archive
contents using an element matching check on a list of strings.