schutzfile: pin osbuild to version ignoring PQC keys
Our service is failing to build RHEL 10 because RHEL 9 doesn't
understand the PQC keys that are now required to build RHEL 10.

As a (temporary?) fix, I implemented
https://github.com/osbuild/osbuild/commit/22d5a83fa8f727ccb411ce315886cab2f95e8fa5

As we are not sure if this is the final solution, we didn't merge it
into main, so this commit actually pins the osbuild version to a
branched version of osbuild. While this is super-ugly, we believe this
is our best shot to restore the service before the holidays.